added modularity to the code
